+++ sdk/storage/azure-storage-files-datalake/inc/azure/storage/files/datalake/datalake_responses.hpp
@@ -15,6 +15,7 @@
 namespace Azure { namespace Storage { namespace Files { namespace DataLake {
 
   class DataLakeServiceClient;
+  class DataLakePathClient;
 
   namespace Models {
 
@@ -246,12 +247,6 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
 
     // DirectoryClient models:
 
-    using SetPathAccessControlListRecursiveSinglePageResult
-        = _detail::PathSetAccessControlRecursiveResult;
-    using UpdatePathAccessControlListRecursiveSinglePageResult
-        = SetPathAccessControlListRecursiveSinglePageResult;
-    using RemovePathAccessControlListRecursiveSinglePageResult
-        = SetPathAccessControlListRecursiveSinglePageResult;
     using CreateDirectoryResult = CreatePathResult;
     using DeleteDirectoryResult = DeletePathResult;
 
@@ -290,5 +285,30 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
     friend class DataLakeDirectoryClient;
     friend class PagedResponse<ListPathsPagedResponse>;
   };
+
+  class SetPathAccessControlListRecursivePagedResponse
+      : public PagedResponse<SetPathAccessControlListRecursivePagedResponse> {
+  public:
+    int32_t NumberOfSuccessfulDirectories = 0;
+    int32_t NumberOfSuccessfulFiles = 0;
+    int32_t NumberOfFailures = 0;
+    std::vector<Models::AclFailedEntry> FailedEntries;
+
+  private:
+    void OnNextPage(const Azure::Core::Context& context);
+
+    std::shared_ptr<DataLakePathClient> m_dataLakePathClient;
+    SetPathAccessControlListRecursiveOptions m_operationOptions;
+    std::vector<Models::Acl> m_acls;
+    _detail::PathSetAccessControlRecursiveMode m_mode;
+
+    friend class DataLakePathClient;
+    friend class PagedResponse<ListPathsPagedResponse>;
+  };
+
+  using UpdatePathAccessControlListRecursivePagedResponse
+      = SetPathAccessControlListRecursivePagedResponse;
+  using RemovePathAccessControlListRecursivePagedResponse
+      = SetPathAccessControlListRecursivePagedResponse;
 
 }}}} // namespace Azure::Storage::Files::DataLake