+++ sdk/storage/azure-storage-files-datalake/test/datalake_directory_client_test.cpp
@@ -382,7 +382,7 @@ namespace Azure { namespace Storage { namespace Test {
     {
       // Set Acls recursive.
       std::vector<Files::DataLake::Models::Acl> acls = GetValidAcls();
-      EXPECT_NO_THROW(rootDirectoryClient.SetAccessControlListRecursiveSinglePage(acls));
+      EXPECT_NO_THROW(rootDirectoryClient.SetAccessControlListRecursive(acls));
       std::vector<Files::DataLake::Models::Acl> resultAcls1;
       std::vector<Files::DataLake::Models::Acl> resultAcls2;
       EXPECT_NO_THROW(resultAcls1 = directoryClient1.GetAccessControlList().Value.Acls);
@@ -409,7 +409,7 @@ namespace Azure { namespace Storage { namespace Test {
       newAcl.Permissions = "rw-";
       std::vector<Files::DataLake::Models::Acl> acls;
       acls.emplace_back(std::move(newAcl));
-      EXPECT_NO_THROW(rootDirectoryClient.UpdateAccessControlListRecursiveSinglePage(acls));
+      EXPECT_NO_THROW(rootDirectoryClient.UpdateAccessControlListRecursive(acls));
       std::vector<Files::DataLake::Models::Acl> resultAcls1;
       std::vector<Files::DataLake::Models::Acl> resultAcls2;
       EXPECT_NO_THROW(resultAcls1 = directoryClient1.GetAccessControlList().Value.Acls);
@@ -486,7 +486,7 @@ namespace Azure { namespace Storage { namespace Test {
       removeAcl.Id = "72a3f86f-271f-439e-b031-25678907d381";
       std::vector<Files::DataLake::Models::Acl> acls;
       acls.emplace_back(std::move(removeAcl));
-      EXPECT_NO_THROW(rootDirectoryClient.RemoveAccessControlListRecursiveSinglePage(acls));
+      EXPECT_NO_THROW(rootDirectoryClient.RemoveAccessControlListRecursive(acls));
       std::vector<Files::DataLake::Models::Acl> resultAcls1;
       std::vector<Files::DataLake::Models::Acl> resultAcls2;
       EXPECT_NO_THROW(resultAcls1 = directoryClient1.GetAccessControlList().Value.Acls);
@@ -562,7 +562,7 @@ namespace Azure { namespace Storage { namespace Test {
         newAcl.Permissions = "rw-";
         acls.emplace_back(std::move(newAcl));
       }
-      (rootDirectoryClient.SetAccessControlListRecursiveSinglePage(acls));
+      (rootDirectoryClient.SetAccessControlListRecursive(acls));
       std::vector<Files::DataLake::Models::Acl> resultAcls1;
       std::vector<Files::DataLake::Models::Acl> resultAcls2;
       EXPECT_NO_THROW(resultAcls1 = directoryClient1.GetAccessControlList().Value.Acls);