+++ sdk/storage/azure-storage-files-datalake/CHANGELOG.md
@@ -10,6 +10,9 @@
   - `DataLakeServiceClient::ListFileSystems`.
   - `DataLakeFileSystemClient::ListPaths`.
   - `DataLakeDirectoryClient::ListPaths`.
+  - `DataLakePathClient::SetAccessControlListRecursive`.
+  - `DataLakePathClient::UpdateAccessControlListRecursive`.
+  - `DataLakePathClient::RemoveAccessControlListRecursive`.
 
 ### Bug Fixes